requirements.txt
flask
flask-cors
google-genai         # official google genai SDK (Gemini)
mysql-connector-python
sqlparse
python-dotenv
